romio: add long long cast to silence printf format warnings
On 64-bit machines, long has the same size as long long. MPICH configures
MPI_OFFSET to long, but ROMIO will configure it to long long. This
raises warnings when we debug print it with the format "%lld" (or %ld).
This commit patches the warnings by casting the argument to "long long".

Many parts of ROMIO code already uses this kind of cast to silence
warnings. This partch merely extends the practice to the rest of code.

I do not enjoy this kind of patching but lack of better alternatives.
Ideally, compilers should offer "auto" format such as `%v` for at least
integer and floating types.

+++ src/mpi/romio/adio/common/ad_aggregate.c
@@ -100,7 +100,7 @@ int ADIOI_Calc_aggregator(ADIO_File fd,
     if (rank_index >= fd->hints->cb_nodes || rank_index < 0) {
         FPRINTF(stderr,
                 "Error in ADIOI_Calc_aggregator(): rank_index(%d) >= fd->hints->cb_nodes (%d) fd_size=%lld off=%lld\n",
-                rank_index, fd->hints->cb_nodes, fd_size, off);
+                rank_index, fd->hints->cb_nodes, (long long) fd_size, (long long) off);
         MPI_Abort(MPI_COMM_WORLD, 1);
     }
 
@@ -405,7 +405,7 @@ void ADIOI_Calc_my_req(ADIO_File fd, ADIO_Offset * offset_list, ADIO_Offset * le
             FPRINTF(stdout, "data needed from %d (count = %d):\n", i, my_req[i].count);
             for (l = 0; l < my_req[i].count; l++) {
                 FPRINTF(stdout, "   off[%d] = %lld, len[%d] = %d\n", l,
-                        my_req[i].offsets[l], l, my_req[i].lens[l]);
+                        (long long) my_req[i].offsets[l], l, (long long) my_req[i].lens[l]);
             }
             FPRINTF(stdout, "buf_idx[%d] = 0x%x\n", i, buf_idx[i]);
         }
